101 playlist invite user management (#103)

+++ supabase/functions/playlists/controller.ts
@@ -1,5 +1,11 @@
 import { Hono } from 'jsr:@hono/hono'
-import { HTTPException } from 'https://deno.land/x/hono@v3.2.3/http-exception.ts'
+import { Context } from 'https://deno.land/x/hono@v3.2.3/mod.ts'
+import {
+  checkPermission,
+  checkPlaylistAccess,
+  PERMISSIONS,
+  getUserRoleInPlaylist
+} from './permissions.ts'
 import { getCurrentUser, getUserToken } from '../auth.ts'
 import {
   createSpotifyPlaylist,
@@ -17,12 +23,16 @@ import {
   deletePlaylistInSupabase,
   addTracksToPlaylistInSupabase,
   deleteTracksFromPlaylistInSupabase,
-  editPlaylistSupabaseById
+  editPlaylistSupabaseById,
+  addUserToPlaylistInSupabase,
+  removeUserFromPlaylistInSupabase
 } from './services/supabase.ts'
 import {
   validateCreatePlaylistPayload,
   validateDeleteTracksPayload,
-  validateEditPlaylistPayload
+  validateEditPlaylistPayload,
+  validateAddUserPayload,
+  validateRemoveUserPayload
 } from './validators.ts';
 
 
@@ -31,26 +41,32 @@ export async function deleteItemsFromPlaylist(c: Context): Promise<any> {
   const body = await c.req.json()
   const { uris } = validateDeleteTracksPayload(body)
 
+  const user = c.get('user')
+
+  await checkPermission(id, user.id, PERMISSIONS.DELETE_SONG)
+
   await deleteTracksFromPlaylistInSupabase(
     id,
     uris
   )
 
   c.status(200)
-  return c.json({ message: 'Tracks deleted successfully' })
 }
 
 export async function addItemsToPlaylist(c: Context): Promise<any> {
   const id = c.req.param('id')
-  const user = c.get('user')
   const body = await c.req.json()
-  const uris = body.uris
-  const position = body.position || 0
+  const { uris } = validateAddTracksPayload(body)
+  const user = c.get('user')
 
-  await addTracksToPlaylistInSupabase(id, uris, user.id)
+  await checkPermission(id, user.id, PERMISSIONS.ADD_SONG)
+
+  await addTracksToPlaylistInSupabase(
+    id,
+    uris
+  )
 
   c.status(201)
-  return c.json({ message: 'Tracks added successfully' })
 }
 
 export async function fetchPlaylistItems(c: Context): Promise<any> {
@@ -58,6 +74,8 @@ export async function fetchPlaylistItems(c: Context): Promise<any> {
   const user = c.get('user')
 
   let playlist = await getSupabasePlaylistById(id)
+
+  checkPlaylistAccess(playlist, user.id)
 
   if (!playlist) {
     c.status(500)
@@ -91,13 +109,36 @@ export async function fetchPlaylistItems(c: Context): Promise<any> {
       };
     });
   }
-  playlist.can_edit = false
-  if (playlist.collaborators.find((collab: any) => collab.id === user.id)) {
-    playlist.can_edit = true
-  }
+  playlist = setPlaylistUserPermissions(playlist, user)
 
+  console.log('Playlist user permissions:', playlist)
   c.status(200)
   return c.json(playlist)
+}
+
+function setPlaylistUserPermissions(playlist: any, user: any) {
+  playlist.user = {}
+  playlist.user.can_edit = false
+  playlist.user.can_invite = false
+  playlist.user.is_following = true
+  playlist.user.role = getUserRoleInPlaylist(playlist, user.id)
+
+  if (playlist.is_collaborative || playlist.collaborators.find((collab: any) => collab.id === user.id)) {
+    playlist.user.can_edit = true
+  }
+  if (playlist.user.role === 'owner' || !playlist.is_private || playlist.collaborators.find((collab: any) => collab.id === user.id)) {
+    playlist.user.can_invite = true
+  }
+
+  const is_owner = playlist.owner.id === user.id
+  const is_member = playlist.members.find((member: any) => member.id === user.id)
+  const is_collaborator = playlist.collaborators.find((collab: any) => collab.id === user.id)
+  if (!is_owner && !is_member && !is_collaborator) {
+    playlist.user.is_following = false
+    playlist.user.can_edit = false
+    playlist.user.can_invite = false
+  }
+  return playlist
 }
 
 
@@ -120,6 +161,7 @@ export async function deletePlaylist(c: Context): Promise<any> {
   const id = c.req.param('id')
   const user = c.get('user')
 
+  await checkPermission(id, user.id, PERMISSIONS.DELETE_PLAYLIST)
   await deletePlaylistInSupabase(id, user.id)
 
   c.status(200)
@@ -128,11 +170,43 @@ export async function deletePlaylist(c: Context): Promise<any> {
 
 export async function updatePlaylist(c: Context): Promise<any> {
   const id = c.req.param('id')
+  const user = c.get('user')
   const body = await c.req.json()
 
   const validatedPayload = validateEditPlaylistPayload(body);
+  await checkPermission(id, user.id, PERMISSIONS.EDIT_PLAYLIST)
 
   await editPlaylistSupabaseById(id, validatedPayload);
   c.status(200)
   return c.json({ message: 'Playlist updated successfully' })
+}
+
+export async function addUserToPlaylist(c: Context): Promise<any> {
+  const id = c.req.param('id')
+  const body = await c.req.json()
+  const user = c.get('user')
+  const { user_id, role } = validateAddUserPayload(body)
+
+  await checkPermission(id, user.id, PERMISSIONS.ADD_USER)
+
+  await addUserToPlaylistInSupabase(id, user_id, role)
+  c.status(200)
+  return c.json({ message: `User added successfully as ${role}` })
+}
+
+export async function removeUserFromPlaylist(c: Context): Promise<any> {
+  const id = c.req.param('id')
+  const body = await c.req.json()
+  const user = c.get('user')
+  const { user_id, role } = validateRemoveUserPayload(body)
+
+  if (user.id !== user_id) {
+    await checkPermission(id, user.id, PERMISSIONS.REMOVE_USER)
+  }
+
+  await removeUserFromPlaylistInSupabase(id, user_id, role)
+
+  console.log('User removed successfully from playlist:', { playlist_id: id, user_id, role });
+  c.status(200)
+  return c.json({ message: 'User removed successfully from playlist' })
 }
